- Adds conversion from Import models into main models - Creates validation functions for amount sign by transaction type, and receipt amount by item amounts - Adds support for "DISCOUNT" imports - Changes amounts from double to BigDecimal

+++ src/main/kotlin/com/liamlime/limefinance/api/model/CurrencyAmountModel.kt
@@ -1,6 +1,8 @@
 package com.liamlime.limefinance.api.model
 
+import java.math.BigDecimal
+
 data class CurrencyAmountModel(
     val currency: CurrencyModel,
-    val amount: Double
+    val amount: BigDecimal
 )
